Move Option to separate packages
