+++ packages/uniforms-material/src/types.ts
@@ -1,0 +1,7 @@
+/** Option type used in SelectField or RadioField */
+export type Option<Value = unknown> = {
+  disabled?: boolean;
+  label: string;
+  key?: string;
+  value: Value;
+};
