+++ packages/uniforms-unstyled/src/SelectField.tsx
@@ -1,6 +1,8 @@
 import xor from 'lodash/xor';
 import React, { Ref } from 'react';
-import { HTMLFieldProps, connectField, filterDOMProps, Option } from 'uniforms';
+import { HTMLFieldProps, connectField, filterDOMProps } from 'uniforms';
+
+import type { Option } from './types';
 
 const base64: (string: string) => string =
   typeof btoa === 'undefined'
