+++ packages/uniforms/src/types.ts
@@ -99,11 +99,3 @@ declare module '.' {
     value: never;
   }
 }
-
-/** Option type used in <SelectField /> */
-export type Option<Value = unknown> = {
-  disabled?: boolean;
-  label: string;
-  key?: string;
-  value: Value;
-};
